chore: renamed file according to convention and added rpcCLient
